Added linear interpolation-based humidity canculation.

+++ humidity.h
@@ -1,4 +1,4 @@
 #define MAXHS 2
 
 void adc_setup (void);
-uint16_t get_humidity (short pin);
+uint8_t get_humidity (short pin);
